+++ packages/styled/example/native/App.tsx
@@ -2,106 +2,61 @@
 import { StyleSheet, Text, View, Pressable } from "react-native";
 
 import { styled } from "@native-base/styled-test";
-// import { Button } from "ui";
-// import { ButtonExpo } from "./button-expo";
 
-const Button = styled(
-  Pressable,
+const Box = styled(
+  View,
   {
     baseStyle: {
       style: {
-        // bg: "$primary.500",
-        px: "$4",
-        py: "$3",
-        // w: 400,}
+        bg: "$red.500",
+        p: "$3",
       },
-      descendants: {
-        _text: {
+      state: {
+        hover: {
           style: {
-            color: "$secondary.500",
+            bg: "$red.300",
           },
         },
       },
     },
     variants: {
-      blueBox: {
-        style: { bg: "$primary.500" },
-        // state: {
-        //   hover: { bg: "$primary.600" },
-        //   focus: { borderWidth: "1", borderColor: "$primary.300" },
-        //   active: { bg: "$primary.800" },
-        // },
-        // p: "$10",
-      },
       greenBox: {
-        bg: "$secondary.400",
-        // px: "$10",
-      },
-    },
-    sizes: {
-      small: {
         style: {
-          w: 100,
-          h: 100,
+          bg: "$secondary.500",
         },
-      },
-      medium: {
-        style: {
-          w: 200,
-          h: 200,
-        },
-      },
-      large: {
-        style: {
-          w: 300,
-          h: 300,
+        state: {
+          hover: {
+            style: {
+              bg: "$primary.600",
+            },
+          },
         },
       },
     },
-    defaultProps: {
-      size: "small",
-      variant: "blueBox",
-    },
-
-    // _text: {
-    //   style: { color: "$primary.100" },
-    //   state: {
-    //     hover: { color: "$red.600" },
-    //     focus: { color: "$primary.300" },
-    //     active: { color: "$primary.900" },
-    //   },
-    // },
   },
-  {
-    descendentStyle: ["_text"],
-    states: ["hover", "focus", "pressed"],
-    aliases: {
-      bg: "backgroundColor",
-    },
-  }
+  {}
 );
+
 export default function App() {
   return (
     <View style={styles.container}>
-      <Text>Open up App.tsx to start working on your app!</Text>
-      {/* <StatusBar style="auto" /> */}
-      {/* <Button
-        onClick={() => {
-          console.log("WORKING");
-        }}
-      /> */}
-      <Button
-        sx={{
-          style: {
-            bg: "$red.800",
-          },
-        }}
-        onPress={() => {
-          console.log("WOORKING BUTTON EXPO");
-        }}
+      <Box
+      // variant="greenBox"
+      // states={{
+      //   hover: false,
+      // }}
+      // sx={{
+      //   state: {
+      //     hover: {
+      //       style: {
+      //         bg: "$red.500",
+      //       },
+      //     },
+      //   },
+      // }}
       >
-        <Text>Hello</Text>
-      </Button>
+        Hello Box
+      </Box>
     </View>
   );
 }