+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @native-base/ui-test
 
+## 1.0.6
+
+### Patch Changes
+
+- Fixed null check
+- Updated dependencies
+  - @native-base/styled-test@1.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- Fixed sx props resolution
+- Updated dependencies
+  - @native-base/styled-test@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes